"boot up" plugins on the fly
