+++ packages/recoil/src/context/index.tsx
@@ -1,4 +1,3 @@
 export * from "./NavEphemeral";
 export * from "./Notifications";
-export * from "./Plugins";
 export * from "./Swap";
